+++ Changelog.md
@@ -5,6 +5,9 @@
 1. Resource deletion CLI.
     * [Pull request 119](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/119)
 
+1. Fetch files
+    * [Pull request 123](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/123)
+
 ### Ignition Fuel Tools 4.0.0 (2019-12-10)
 
 1. Model upload CLI, resource metadata CLI, depend on ign-msgs5.